requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
pydantic==2.5.1
python-dotenv==1.0.0
prometheus-flask-exporter==0.22.4
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
mypy==1.7.0
flake8==6.1.0